requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
cairocffi==1.7.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.8.0
defusedxml==0.7.1
freetype-py==2.5.1
google-auth==2.42.1
google-genai==1.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageHash==4.3.2
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.3.4
pillow==12.0.0
pillow_heif==1.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyWavelets==1.9.0
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
rlPyCairo==0.4.0
rsa==4.9.1
scipy==1.16.3
sniffio==1.3.1
tenacity==9.1.2
tinycss2==1.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
webencodings==0.5.1
websockets==15.0.1
